+++ CHANGELOG.md
@@ -4,12 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-# 0.0.7 - 2025-01-20
+# 0.0.7 - 2025-01-21
+
+### Added
+
+- Completion for Includes
+- Diagnostic for Include when server is unable to resolve
 
 ### Fixed
 
 - Disable `diagnostics.refresh` if Client does not support this. [Fixed by [topisani](https://github.com/topisani)]
 - Invalid uri format in diagnostics when clearing workspace diagnostics `diagnostics.refresh`. [Fixed by [topisani](https://github.com/topisani)]
+- Server crash if import is a directory
+- Report diagnostics issues from CPrePorcessor
 
 # 0.0.6 - 2025-01-18